Health PowerUP/Trap Script
I am in the process of figuring out spawning the health buff, once i figure that out i will do the trap as well. In the meantime, i delelted the trap asset so that nothing is in the way of the health buff script to help with debugging. I also removed all the other unnecessary powerup scripts and the previous original spawn script.