Recursive Error + Treasure Scoring System
Created a scoring system that will search and sort treasure items by value of the treasure in the treasure chest, haven't been able to test code out yet because as soon as I connected it, the game system crashed due to a 'Stack overflow' error, meaning there is an infinte recursive case somewhere, most likelly come from connecting the scoring system to the treasure to the saving system all at once. Instead, will work on treasure system first, then connect it to a score board that updates on real time, then focus on saving it and loading it. Next goal is the settings for a new game, quitting and starting a game again.