TreasureChest + Saving System + Speed Buff
I fixed the spawning issue of the speed buff, it was applying speed buff affect but it wasn't spawning, I had to change the order of the script and box collider for it to work, but now its spawning so thats good. I wanted the chest to not by a trigger collider after getting the item in the chest so I added the colilder component to the code and made it so that after the open chest function was called, in the end it changed the value of the isTrigger to false sot that the player can collide with it rather than passing right from under. Next I made sure that the health bar could update by making it use UnityEngine.UI which it was not which is why when saved, the healthbar wasn't updating and then I made it so that when saving and loading, the current score was also included.