Main Menu Addition
In the process of creating a main menu that will have 3 options, loading the current saved game, making a new blank maze game, or just quitting the application. Later, I'll link the player's health so that when it dies, it goes back the menu section without saving, but it'll still load previous saved scene. I might add music later, but Im not sure, and I have to also add portals, a few more additions must be made basically...