Major Changes for Finalized MVP
Cleaned up files, removed ones that I don't use like trap trigger and maze initialization scripts. Was in the process of incorporating multiple scenes to have menu screens when pressing options or when the game starts, but due to major errors, had to remove the scenes and scripts related to it (optionmenu and mainmenu script). Cleaned up assets tab to put every script in its place. Added portal images, will make scripts for it later. As well, finally added top-down movement for the character, rather than having a square, it now has updated movements where when an arrow key is pressed, its respective sprite is there, so left key has character facing left, etc, for better image. Trying to fix the save system, health errors. The system crashes when there is no data to access from in loaddata method, although the try/catch error statements are there, so have to figure out why that happens. The health bar doesn't have a max or minimum value either, it does on the health bar, but when looking at the players health value, it can be above 100 and below 0 even though the health bar doesn't show it. The if/else statements to handle that are there but it doesn't work so have to figure that out too. MVP looks good apart from all that.