Minor Insignificant Change
I just changed the anchor of the save/load game buttons to top right as when I maximized the game screen for my screenshots it was in the center, you wouldn't know that it wasn't anchored properly before if it wasn't maximized. I just changed it for presentation purposes as its weird to have the buttons floating in the screen.